コンポーネントの分割 pull/4#discussion_r1799414319への対応
