+++ src/components/DisplayErrors.tsx
@@ -1,0 +1,34 @@
+function DisplayErrors(errorMessages: string){
+  if(errorMessages == ""){ // エラーが存在しない場合
+    return false;
+  }
+  // 受け取ったエラーメッセージを表示用に変える。
+  // 指定しているもの以外のエラーに関しては配列に含めたくないのでmapではなくforEachを使用する
+  const displayedErrorMessages: Array<string> =[];
+  errorMessages.split("\n").forEach((errorMessage) => {
+    if(errorMessage.includes("Name")){
+      displayedErrorMessages.push("nameが空欄です。");
+    }else if(errorMessage.includes("Email")){
+      displayedErrorMessages.push("Emailに誤りがあるため確認してください。");
+    }else if(errorMessage.includes("年齢")){
+      displayedErrorMessages.push("年齢を選択してください。");
+    }
+  })
+  // エラーを表示する
+  return (
+    <div className="errorMessage">
+      {(displayedErrorMessages.length == 0)? <></>:(
+        <>
+          <p>エラーが{displayedErrorMessages.length}個あります。</p>
+          <ul>
+            {displayedErrorMessages.map((msg) => 
+              <li key={msg}>{msg}</li>
+            )}
+          </ul>
+        </>
+      )}
+    </div> 
+  );
+}
+
+export default DisplayErrors;
